refactor (back): remodelando a arqutetura para o nest
