+++ back/src/factories/authentication/makeAuthenticationMiddleware.ts
@@ -1,0 +1,6 @@
+import { AuthenticationMiddleware } from '../../app/shared/middlewares/AuthenticationMiddleware';
+
+
+export function makeAuthenticationMiddleware() {
+  return new AuthenticationMiddleware();
+}
