--- back/src/factories/makeAuthenticationMiddleware.ts
@@ -1,5 +1,0 @@
-import { AuthenticationMiddleware } from '../app/middlewares/AuthenticationMiddleware';
-
-export function makeAuthenticationMiddleware() {
-  return new AuthenticationMiddleware();
-}
