--- back/src/factories/makeSignInUseCase.ts
@@ -1,8 +1,0 @@
-import { AccountRepository } from '../app/Repositories/AccountRepository';
-import { SignInUseCase } from '../app/useCases/SignInUseCase';
-
-export function makeSignInUseCase() {
-  const accountRepository = new AccountRepository();
-
-  return new SignInUseCase(accountRepository);
-}
