--- back/src/factories/mekeEditAccountController.ts
@@ -1,5 +1,0 @@
-import { EditAccountController } from '../app/controllers/editAccountController';
-
-export function makeEditAccountController() {
-  return new EditAccountController();
-}
