+++ back/src/factories/signIn/makeSignInUseCase.ts
@@ -1,0 +1,8 @@
+import { SignInUseCase } from '../../app/modules/users/signIn/SignInUseCase';
+import { UsersRepository } from '../../app/modules/users/UsersRepository';
+
+export function makeSignInUseCase() {
+  const usersRepository = new UsersRepository();
+
+  return new SignInUseCase(usersRepository);
+}
